Merged all the projects into one.
Took the new version of the prototype that Adam and Sam created  with the new improved camera  and imported the Main Menu that i have created in the past  together with the new Menu that i have created for the game level.
Have also created a subfolders in the scripts and organized them acroding to each scene that they are used.