Added new Menu Scene
Added the new temporary version of the meny system.

The player will not inhabit a small "waiting room" style system where the menus are around the player, the player can move around a small limited space and interact with the menus using raycasting (Panos is working on making the raycasting work properly, either that or we'll use colliders and the player will put there hand inside a "start game" zone to start the game instead of aiming and clicking.

The current temp zone for the waiting room is the test level (since we wont be using that in the final game its a nice homage to what we're used to), and the placeholder textures are put in place of where the final panels for the menus will do.

We thought michael might want to be in charge of decorating the area a bit better to make the menu look a bit more impressive, since its just playing around with assets that are already present in teh build.